Rename value filter classes and split into two separate headers
