+++ libtiledbsoma/src/soma/coordinate_value_filters.cc
@@ -1,5 +1,5 @@
 /**
- * @file   soma_query_condition.cc
+ * @file   coordinate_value_filters.cc
  *
  * @section LICENSE
  *
@@ -8,34 +8,19 @@
  *
  * @section DESCRIPTION
  *
- * This file defines helper functions and classes for using query conditions in SOMA.
+ * This file defines the CoordinateValueFilters class for construction query conditions from column selections.
  */
 
-#include "soma_query_condition.h"
+#include "coordinate_value_filters.h"
 
 #include <numeric>
+#include "../tiledb_adapter/value_filter.h"
 #include "../utils/logger.h"
 
 namespace tiledbsoma {
 using namespace tiledb;
 
-/**************************************
- * SOMAValueFilter
-**************************************/
-
-SOMAValueFilter::SOMAValueFilter(const QueryCondition& qc)
-    : qc_{qc} {
-}
-
-SOMAValueFilter::SOMAValueFilter(QueryCondition&& qc)
-    : qc_{qc} {
-}
-
-/**************************************
- * CoordinateValueFilter
-**************************************/
-
-CoordinateValueFilter::CoordinateValueFilter(
+CoordinateValueFilters::CoordinateValueFilters(
     std::shared_ptr<Array> array,
     std::shared_ptr<SOMAContext> ctx,
     std::vector<std::shared_ptr<SOMAColumn>> index_columns,
@@ -47,15 +32,15 @@ CoordinateValueFilter::CoordinateValueFilter(
     , coord_qc_(index_columns_.size()) {
 }
 
-bool CoordinateValueFilter::is_initialized() const {
+bool CoordinateValueFilters::is_initialized() const {
     return std::any_of(coord_qc_.cbegin(), coord_qc_.cend(), [](auto qc) { return qc.is_initialized(); });
 }
 
-SOMAValueFilter CoordinateValueFilter::get_value_filter() const {
-    return std::reduce(coord_qc_.cbegin(), coord_qc_.cend(), SOMAValueFilter(), [](const auto& qc1, const auto& qc2) {
+ValueFilter CoordinateValueFilters::get_value_filter() const {
+    return std::reduce(coord_qc_.cbegin(), coord_qc_.cend(), ValueFilter(), [](const auto& qc1, const auto& qc2) {
         if (qc1.is_initialized()) {
             if (qc2.is_initialized()) {
-                return SOMAValueFilter(qc1.query_condition().combine(qc2.query_condition(), TILEDB_AND));
+                return ValueFilter(qc1.query_condition().combine(qc2.query_condition(), TILEDB_AND));
             }
             return qc1;
         }
@@ -63,14 +48,14 @@ SOMAValueFilter CoordinateValueFilter::get_value_filter() const {
     });
 }
 
-CoordinateValueFilter& CoordinateValueFilter::add_coordinate_query_condition(int64_t index, SOMAValueFilter&& qc) {
+CoordinateValueFilters& CoordinateValueFilters::add_coordinate_query_condition(int64_t index, ValueFilter&& qc) {
     if (!qc.is_initialized()) {
         // No-op.
         return *this;
     }
     auto& current = coord_qc_[index];
     if (current.is_initialized()) {
-        coord_qc_[index] = SOMAValueFilter(current.query_condition().combine(qc.query_condition(), TILEDB_OR));
+        coord_qc_[index] = ValueFilter(current.query_condition().combine(qc.query_condition(), TILEDB_OR));
     } else {
         coord_qc_[index] = qc;
     }
@@ -78,7 +63,7 @@ CoordinateValueFilter& CoordinateValueFilter::add_coordinate_query_condition(int
 }
 
 /**Throws an error if using a string on a non-string column. */
-void CoordinateValueFilter::validate_string_column(std::shared_ptr<SOMAColumn> column) const {
+void CoordinateValueFilters::validate_string_column(std::shared_ptr<SOMAColumn> column) const {
     if (column->domain_type().value() != TILEDB_STRING_ASCII) {
         std::stringstream ss;
         auto tiledb_type = tiledb::impl::type_to_str(column->domain_type().value());
