--- libtiledbsoma/src/tiledb_adapter/soma_query_condition.h
@@ -1,326 +1,0 @@
-/**
- * @file   soma_query_condition.h
- *
- * @section LICENSE
- *
- * Licensed under the MIT License.
- * Copyright (c) TileDB, Inc. and The Chan Zuckerberg Initiative Foundation
- *
- * @section DESCRIPTION
- *
- * This file defines helper functions and classes for using query conditions in SOMA.
- */
-
-#ifndef TILEDBSOMA_TILEDB_ADAPTER_H
-#define TILEDBSOMA_TILEDB_ADAPTER_H
-
-#include <memory>
-#include <optional>
-#include <span>
-#include <sstream>
-#include <vector>
-
-#include <tiledb/tiledb.h>
-#include <tiledb/tiledb>
-
-#include "../common/soma_column_selection.h"
-#include "../soma/soma_column.h"
-#include "../soma/soma_context.h"
-#include "../utils/common.h"
-
-namespace tiledbsoma {
-using namespace tiledb;
-
-class SOMAValueFilter {
-   public:
-    SOMAValueFilter() = default;
-    SOMAValueFilter(SOMAValueFilter&& other) = default;
-    SOMAValueFilter(const SOMAValueFilter& other) = default;
-    ~SOMAValueFilter() = default;
-
-    SOMAValueFilter& operator=(const SOMAValueFilter& other) = default;
-    SOMAValueFilter& operator=(SOMAValueFilter&& other) = default;
-
-    /**
-     * Create a SOMAValueFilter from an initialized TileDB query condition.
-     *
-     * Important: The TileDB QueryCondtion must be initialized.
-     *
-     * @param qc The initialized TileDB query condition.
-     */
-    SOMAValueFilter(const QueryCondition& qc);
-
-    /**
-     * Create a SOMAValueFilter from an initialized TileDB query condition.
-     *
-     * Important: The TileDB QueryCondtion must be initialized.
-     *
-     * @param qc The initialized TileDB query condition.
-     */
-    SOMAValueFilter(QueryCondition&& qc);
-
-    /**
-     * Create a query condition from a slice.
-     *
-     * @tparam The type of the element the query condition applies to.
-     * @param column_name The name of the element the query condition applies to.
-     * @param start_value The start of the range (inclusive).
-     * @param stop_value The end of the range (inclusive).
-     */
-    template <typename T>
-    static SOMAValueFilter create_from_slice(
-        const Context& ctx, const std::string& column_name, const SOMASliceSelection<T>& slice) {
-        if (slice.start.has_value() && slice.stop.has_value()) {
-            return SOMAValueFilter(
-                QueryCondition::create(ctx, column_name, slice.start.value(), TILEDB_GE)
-                    .combine(QueryCondition::create(ctx, column_name, slice.stop.value(), TILEDB_LE), TILEDB_AND));
-        }
-        if (slice.start.has_value()) {
-            return SOMAValueFilter(QueryCondition::create(ctx, column_name, slice.start.value(), TILEDB_GE));
-        }
-        if (slice.stop.has_value()) {
-            return SOMAValueFilter(QueryCondition::create(ctx, column_name, slice.stop.value(), TILEDB_LE));
-        }
-        return SOMAValueFilter();
-    }
-
-    /**
-     * Create a query condition from a series of points on a range.
-     *
-     * @tparam The type of the element the query condition applies to.
-     * @param column_name The name of the element the query condition applies to.
-     * @param values The values of the points
-     */
-    template <typename T, typename = std::enable_if<!std::is_same_v<T, std::string>>>
-    static SOMAValueFilter create_from_points(
-        const Context& ctx, const std::string& column_name, SOMAPointSelection<T> values) {
-        if (values.points.empty()) {
-            // Use sstream beacuse we don't want to include fmt directly in external header.
-            std::stringstream ss;
-            ss << "Cannot set coordinates on column '" << column_name << "'. No coordinates provided.";
-            throw std::invalid_argument(ss.str());
-        }
-        if constexpr (std::is_same_v<T, std::string>) {
-            // Using C API because C++ API only supports std::vector, not std::span.
-            uint64_t data_size = 0;
-            for (auto& val : values.points) {
-                data_size += val.size();
-            }
-            std::vector<uint8_t> data(data_size);
-            std::vector<uint64_t> offsets{};
-            uint64_t curr_offset = 0;
-            for (auto& val : values.points) {
-                offsets.push_back(curr_offset);
-                memcpy(data.data() + curr_offset, val.data(), val.size());
-                curr_offset += val.size();
-            }
-            tiledb_query_condition_t* qc;
-            ctx.handle_error(tiledb_query_condition_alloc_set_membership(
-                ctx.ptr().get(),
-                column_name.c_str(),
-                data.data(),
-                data.size(),
-                offsets.data(),
-                offsets.size() * sizeof(uint64_t),
-                TILEDB_IN,
-                &qc));
-            return QueryCondition(ctx, qc);
-        } else {
-            // Using C API because C++ API only supports std::vector, not std::span.
-            std::vector<uint64_t> offsets(values.points.size());
-            for (size_t index = 0; index < values.points.size(); ++index) {
-                offsets[index] = index * sizeof(T);
-            }
-            tiledb_query_condition_t* qc;
-            ctx.handle_error(tiledb_query_condition_alloc_set_membership(
-                ctx.ptr().get(),
-                column_name.c_str(),
-                values.points.data(),
-                values.points.size() * sizeof(T),
-                offsets.data(),
-                offsets.size() * sizeof(uint64_t),
-                TILEDB_IN,
-                &qc));
-            return QueryCondition(ctx, qc);
-        }
-    }
-
-    /**
-     * Return if the query condition is initialized.
-     */
-    inline bool is_initialized() const {
-        return qc_.has_value();
-    }
-
-    /**
-     * Return internal TileDB query condition.
-     */
-    inline const QueryCondition& query_condition() const {
-        if (!qc_.has_value()) {
-            throw TileDBSOMAError("Internal error: Cannot return query condition. Query condition is not initialized.");
-        }
-        return qc_.value();
-    }
-
-   private:
-    /** Query condition for coordinates. */
-    std::optional<QueryCondition> qc_;
-};
-
-class CoordinateValueFilter {
-   public:
-    CoordinateValueFilter() = delete;
-    CoordinateValueFilter(
-        std::shared_ptr<Array> array,
-        std::shared_ptr<SOMAContext> ctx,
-        std::vector<std::shared_ptr<SOMAColumn>> index_columns,
-        Domainish domain_kind);
-    CoordinateValueFilter(CoordinateValueFilter&& other) = default;
-    CoordinateValueFilter(const CoordinateValueFilter& other) = default;
-    ~CoordinateValueFilter() = default;
-
-    /**
-     * @brief Add a constraint to a column.
-     *
-     * The query condition will select on elements of the TileDB array that include coordinates
-     * within the selected range for the column (inclusive of end points). The slice must intersect
-     * the domain.
-     *
-     * @tparam T The type of the column the query condition is applied to.
-     * @param col_index The index of the column.
-     * @param slice The coordinate selection to apply to the columns.
-     */
-    template <typename T>
-    CoordinateValueFilter& add_slice(int64_t col_index, const SOMASliceSelection<T>& selection) {
-        auto col = index_columns_[col_index];
-        if constexpr (std::is_same_v<T, std::string>) {
-            validate_string_column(col);
-        } else {
-            try {
-                impl::type_check<T>(col->domain_type().value());
-            } catch (const TileDBError& e) {
-                throw std::invalid_argument("Invalid type.");  // TODO: Add error message before merging.
-            }
-            auto domain = col->domain_slot<T>(*ctx_, *array_, domain_kind_);
-            if (!selection.has_overlap(domain)) {
-                // Use sstream beacuse we don't want to include fmt directly in external header.
-                std::stringstream ss;
-                ss << "Non-overlapping slice [" << selection.start.value_or(domain.first) << ", "
-                   << selection.stop.value_or(domain.second) << "] on column '" << col->name()
-                   << "'. Slice must overlap the current column domain [" << domain.first << ", " << domain.second
-                   << "].";
-                throw std::out_of_range(ss.str());
-            }
-        }
-        add_coordinate_query_condition(
-            col_index, SOMAValueFilter::create_from_slice<T>(*ctx_->tiledb_ctx(), col->name(), selection));
-
-        return *this;
-    }
-
-    /**
-     * @brief Add a constraint to a column.
-     *
-     * For a series of points: The query condition will select on elements of the TileDB array that 
-     * include coordinates that equal one of the provided elements for the specified attribute or
-     * dimension. This method is not recommended for floating-point components. All points must be
-     * inside the domain.
-     *
-     * @tparam T The type of the column the query condition is applied to.
-     * @param col_index The index of the column.
-     * @param points The coordinate selection to apply to the columns.
-     */
-    template <typename T>
-    CoordinateValueFilter& add_points(int64_t col_index, SOMAPointSelection<T> selection) {
-        auto col = index_columns_[col_index];
-        if constexpr (std::is_same_v<T, std::string>) {
-            validate_string_column(col);
-        } else {
-            try {
-                impl::type_check<T>(col->domain_type().value());
-            } catch (const TileDBError& e) {
-                throw std::invalid_argument("Invalid type.");  // TODO: Add error message before merging.
-            }
-            auto domain = col->domain_slot<T>(*ctx_, *array_, domain_kind_);
-            if (!selection.is_subset(domain)) {
-                // Use sstream beacuse we don't want to include fmt directly in external header.
-                std::stringstream ss;
-                ss << "Out-of-bounds coordinates found on column '" << col->name()
-                   << "'. Coordinates must be inside column domain [" << domain.first << ", " << domain.second << "].";
-                throw std::out_of_range(ss.str());
-            }
-        }
-        return add_coordinate_query_condition(
-            col_index, SOMAValueFilter::create_from_points<T>(*ctx_->tiledb_ctx(), col->name(), selection));
-    }
-
-    /**
-     * @brief Add a constraint to a column.
-     *
-     * For a slice constraint: the query condition will select on elements of the TileDB array that 
-     * include coordinates within the selected range for the column (inclusive of end points). The
-     * slice must intersect the domain.
-     *
-     * For a series of points: The query condition will select on elements of the TileDB array that
-     * include coordinates that equal one of the provided elements for the specified attribute or
-     * dimension. This method is not recommended for floating-point components. All points must be
-     * inside the domain.
-     *
-     * For monostate: no constraint is applied.
-     *
-     * @tparam T The type of the column the query condition is applied to.
-     * @param col_index The index of the column.
-     * @param selection The coordinate selection to apply to the columns.
-     */
-    template <typename T>
-    CoordinateValueFilter& add_column_selection(int64_t col_index, SOMAColumnSelection<T> selection) {
-        std::visit(
-            [&](auto&& val) {
-                using S = std::decay_t<decltype(val)>;
-                if constexpr (std::is_same_v<S, SOMASliceSelection<T>>) {
-                    add_slice<T>(col_index, val);
-                } else if constexpr (std::is_same_v<S, SOMAPointSelection<T>>) {
-                    add_points<T>(col_index, val);
-                }
-                // Otherwise monostate: do nothing.
-            },
-            selection);
-        return *this;
-    }
-
-    /**
-     * Returns `true` if at least one query condition is set.
-     */
-    bool is_initialized() const;
-
-    /**
-     * Returns a SOMA query condition that queries for the currently set coordinates.
-     */
-    SOMAValueFilter get_value_filter() const;
-
-   private:
-    /**Add a query condition to a coordinate. */
-    CoordinateValueFilter& add_coordinate_query_condition(int64_t index, SOMAValueFilter&& qc);
-
-    /**Throws an error if using a string on a non-string column. */
-    void validate_string_column(std::shared_ptr<SOMAColumn> column) const;
-
-    /** TileDB context required for creating query conditions. */
-    std::shared_ptr<SOMAContext> ctx_;
-
-    /** TileDB Array (only included because it's required to get domain from a column). */
-    std::shared_ptr<Array> array_;
-
-    /** Names of the dimensions of the underlying TileDB Array. */
-    std::vector<std::shared_ptr<SOMAColumn>> index_columns_;
-
-    /** The type of domain stored in the array. */
-    Domainish domain_kind_;
-
-    /** Query condition for coordinates. */
-    std::vector<SOMAValueFilter> coord_qc_;
-};
-
-}  // namespace tiledbsoma
-
-#endif
