+++ libtiledbsoma/src/tiledbsoma/tiledbsoma
@@ -28,6 +28,7 @@
 #include "soma/logger_public.h"
 #include "soma/soma_context.h"
 #include "soma/managed_query.h"
+#include "soma/coordinate_value_filters.h"
 #include "soma/array_buffers.h"
 #include "soma/column_buffer.h"
 #include "soma/soma_coordinates.h"
